pivots and get func
button.name returns a string of the button's name; I renamed the buttons to button1 button2 etc and took the last character of the string so it returns 1 2 3 4 or 5.
I wrote in a get function; there should be some separate script that would call that function for all the dialogue boxes but idk where to have that script be (on phil? or smthn else)
a note on int vs char: the last character is a char. if i tried casting it to int like (int)"(the string '1' '2' '3' etc) it gave me numbers like 51, 52, etc, which makes me think its ascii or smthn. I converted it using the function getNumericValue that C# has, so now it returns a double (idk if theres one that returns it as an int but i think its not too important)